Add method to FilePath to create a Config if missing
The DeploymentServer needs to be able to access to Config before it
is set by BaseCommand via the FilePath. To facilitate this, a method
has been added which will allow FilePath to create a new Config.

This is likely to be the only use case where a Config is used by hasn't
been already cached.