Update Staging to use a class method as the validator
The validator is no longer a initialized class. Instead it can be any
class that has `validate` method which takes the file content.

Also the string needs to be converted to a ruby Object using constantize.
Error handling has been added to the validation. If an error is raised,
then it will be interpreted as a failure. The raised error message is
included in the warning.