Remove the obsolete DomainTemplateRenderer class
The domain templates are no longer rendered at the same time every
time. Instead the server.yaml file is read directly; the genders file
is rendered to host on configure; and dns templates are rendered on
template. See previous commits for further details of these changes.

This means the DomainTemplateRenderer is no longer required. Note
the managed file are now handled by sync. It also means the
`configure rerender` command needed to be removed. It was solely
a wrapper around the DTR. A similar set of functionality might be
required in the future, however this would need to be implemented
differently.

Some of the tests needed to be updated to fix require errors.