Create DNS::Hosts method and render on template
A DNS type has been created for the Hosts template. It is based off
Named. However it is quite a bit simpler then Named as it is a single
template.

The rendering of the DNS templates happen on every `template` command.
To facilitate the use of a shared staging cache, the BuildMethod needed
to be updated. The render staging templates now takes a templator object
which contains the staging cache.