Replace DHCP command with rendering a file on template
Instead of specifically running a command to rerender the DHCP file,
a RenderMethod has been made for DHCP. The file will automatically
be rerendered each time the `template` command is ran.

This means it will now be picked up by the template, stage, sync cycle.

Currently the code to restart services is not being ran. This will
eventually occur at the end of sync