Convert staging manifest files list to a hash
Previously the files list was an array. This means if the same file
is rendered multiple times, it will have multiple entries. However
only the first entry will be synced and the others will error.

The fix is to build the manifest on a hash with the sync location
being the key. Also as there is a 1-1 mapping between sync locations
and the staging directory, the staging path does not need to be
saved.