Split sync_files command into a iterator function
The sync_files command is not well suited to the Staging class as it
coupled to the Sync command. Instead the underlining iterator has been
broken out of the code.

delete_file_if will run a block for each staged file. If the block
returns a truthy file, it is deleted from the staging manifest and
directory.

This allows the Sync specific code to sit on top as a block. As all
file commands currently need its content, delete_file_if will auto
load it into the returned structure. Thus use the iterator means the
Managed files are handled implicitly.

Also added a restart_service field to the manifest. As files are synced
the services list needs to be updated. This is how metalware will know
what to restart at the end.