Cache and restart services during the Sync command
As the Sync command moves files into place, it updates the services
list in the manifest. This will keep the record of which services need
to be restarted at the end.

If a service restarts correctly it will be removed from the list,
otherwise it remains in the manifest.

Note the manifest is saved before the services are restarted. This is
to insure that any exceptions do not cause the file changes to be lost.

The manifest will automatically be saved again at the end of the update
block.