Create a scope method that defaults to Domain
Currently the scope used by metalware is not enforced. Instead it is
set by the parameters passed as the dynamic namespace. Ideally the
behaviour would be standardised to make setting the scope more intuitive.

The first part of the spec is it must default to the domain namespace.
This ensures that something is always in scope.