Allow the Node or Group to be set as the scope
When a :node xor :group key is passed into the dynamic namespace, it
is set as the scope.

Please note the class is of the key is not currently correct. However
the scope should only be set by Node or Group class themselves, which
will use 'self'. Therefore the class of the variable (:node or .:group)
is not anticipated to be a problem.