Delegate the config, and answer to the in scope object
Instead of passing in the config and answer hashes into the dynamic
namespace, they are delegate to the scope.

This ensures the config remains in sync with the scope. Also it means
the config and answers are no longer passed into alces.

Because the scope is now set, the funky code that redirects to
alces.domain.render_erb_template is no longer required.