Update and save the internal state in GroupCache
Previously the new state was determined and then passed to the save
method for the GroupCache. However this causes issues for the
implementation of the orphan node.

Instead the internal state of groups hash should be updated. Then
the save method can directly pull the data it needs to save.

Also use a new cache in the group_spec as it is more reliable