Ensure the local Configurator is used instead of node
Moving forward in Metalware, it is becoming increasingly more common
to treat the local node as a regular node if possible. This is because
the local node appears in `alces.nodes.each`.

Because Configurator makes a distinction between the two, it should
be responsible for making sure the local method is called and not
the caller.