Use the Alces Namespace to get the default answers
Previously the Configurator would merge its own answer files to obtain
the previous default. This was so it could tell the difference between
higher level defaults that flow through AND it's own saved answer.

However the merging of hashes is done by HashMerger to ensure
consistency. It does mean some additional logic is required to determine
if the answer needs to be saved.

The answer does not get saved the default came from a higher level
config (aka there is no old answer). However it does need to be saved
if the default is the old answer as to prevent it being lost.

The tests have been simplified as they no longer need to test the
merging of the hashes.

As the Alces namespace is being used during the Configurator, it
can not bee used to render Genders. This is because the answer files
have been changed and thus the cache is stale. This means a new
namespace must be created for rendering the genders file.