Fix scope bug when the dynamic namespace isn't set
Internally within metalware, there are cases where the namespace is
used outside of templating. In these cases, there is no dynamic namespace
within the Alces namespace.

However the scope command relies on the dynamic namespace being set
in order to default to the domain scope. The fix is to use an empty
struct if the dynamic namespace hasn't been set.

This will allow `alces` to act as a shorthand to alces.domain when
using the various metalware tools. The primary example is FilePath
that uses the <namespace>.config method to find templates. This
relies on alces.config being equivalent to alces.domain.config.

This isn't an issue during rendering as the dynamic namespace is
always set.

The remaining file changes are to do with the spoofing of the redirect.
During the transition to the new namespace, there was a need to prevent
domain from being the default. This is no longer the case as the code
base has been improved since then. Thus the config flag and utility
to turn of the redirect have been removed.