Add missing nodes to the orphan group
When an orphan node is being configured, it wont exist in the nodes
list in the namespace. When this happens, the node will be added as
an orphan. The defaults will be those specified at the orphan group
level.

As the local node now belongs to the orphan group, the same thing
will happen to it. However the warning will not be displayed for the
local node as it will always happen.