Change the local group to the orphan group
The local group is being deprecated, instead the local node is going
to be made in the orphan group. The only thing that previously put
the local node in the local group is the fact it appears in the
genders file.

Moving forward, `configure local` will trigger the local node to be
puts into the orphan group