Create BuildEvent object to handle the build hooks
The code that gets ran on each hook is stored with in the BuildMethods.
The purpose of the BuildEvent however is to abstract away the steps
involved in the build.

That way Build can control the flow of the build, BuildMethods control
the code that is ran, and BuildEvent translates between the two.

This commit implements the command to run all the start hooks and
to check if the build has finished.

All the hooks will be ran in there own thread. This way metalware
can continue checking for built nodes whilst they run.

The build will only register complete once the internal list of nodes
is empty. Eventually built nodes will be removed from the list. Also
a build can not finish until all the hooks have died. NOTE: sleeping
threads will not be killed.

A kill_threads method has also been included to end all the hooks