Wait for threads to join when killing them
The kill_threads method will hang until all the build_threads have
been killed. In practice this should be a fast operation.

It does have the possibility to hang forever, but this would be
indicative of a greater problem. This case will not be handled in
BuildEvent as it will be handled in the Build commands flow control.