Integrate BuildEvent into the build command
The loop used to build the nodes can be simplified a bit using the
new BuildEvent. There is still a degree of coupling between the two
however this can be addressed at a future date.

NOTE: The GUI bit is probably broken as the loop now ends when the
build is complete. But the GUI is broken anyway so this can also be
addressed latter.