Fix Configure bug when pulling answers for new group
The `configure group` command was broken for previously unconfigured
groups (not super useful). This was because it couldn't find the
answers for the group (which contain the defaults) as it didn't
appear in the namespace. It does appear in the namespace as it is
unconfigured.

Instead a new group namespace for that group needs to be created as
it will load the answers. A similar thing needed to happen for the
Node. (Previously nodes used the group answers, however this doesn't
work as groups have different questions).

The configure command for group has been simplified. Also the GroupCache
needs to expose the method for getting the next group index. Also
the config input to GroupCache needs to be removed.