Merger the answer hash for unconfigured orphan nodes
In order to obtain the defaults answers, the hash_merger must obtain
a list of groups the nodes belong to. This breaks for orphan nodes as
they do not appear in the genders file.

Instead, the orphan group must be passed in explicitly in the event
of an error.

Also broke down the code in Configurator for readability.