Update Validation spec to properly load error file
The error file needed to be loaded into the faked filesystem. Exactly
why this wasn't an issue before is unknown. However cloning the file
makes debugging easy for the future.

Also for some other unknown reason, dry validation requires both the
QuestionSchema and the DependantiSchema to use the same config.