Flag the success? status instead of calculating it
Instead of traversing the tree to determine if the validation was
successful, the status can be flag. As each question is validated,
any failed validations will set a failed flag. The success? method
then checks if the flag has been set and returns the opposite.

Also changed the location a failed validation is raised. There is no
reason to have the code spread over two methods.