Print questions that failed validation
As the questions are now based on a tree, the error results can be
printed by traversing the tree.

To assist the successful nodes are removed from the tree. Please note
only leaf nodes are removed as deleting branches would deleted failed
leaves.

The rubytree gem needed to be updated so the pruning function could
use a postordered tree traversal.