Migrate the old Schema checks as a new schema
The QuestionSchema has been split into two parts, a wrapper that
will check if the default matches the type. It is easier to do this
in a wrapper as the entire hash can be passed to a custom predicate.

Then there is a second schema that checks individual fields,
QuestionFieldSchema. This schema checks that the individual fields
are correct.