Fix bug where questions didn't validate dependents
The Validation using the DependentSchema in a question needs to occur
in the QuestionSchema and not the QuestionFieldSchema. This is
because the results of the DependentSchema need to be chained into
the greater results. This can only occur in the QuestionSchema for
reasons. (Dry::Validation is a pain and needs to be used certain ways)

Also removed the results.nil? check when printing the errors. It is
no longer required as all nodes now have a results field.