Remove the requirement that each section is specified
Each section is no longer required. If a section is not specified,
it is assumed there are now questions. This will also happen if the
section is explicitly set to nil or false.

If it is set any other value; then the requirement that it is an
array (including empty array) is enforced.