Load and validate configure.yaml when initializing
The Validation::Configure object is initialized with either a
hash or YAML, which can not be changed once the object is created.

This means the internal tree objected can be created and validated
on new as well. There are no time savings either way as the tree
object needs to be created regardless.