Update answer validation to use changes in configure
Now that the configure questions are built on a tree, the answers
are no longer in sync. The answers are expecting a hash based on the
identifiers of the questions.

This is still possible as each sub question should still have an
unique identifier. This way the answer validation can be kept the
same with minimal changes.

Instead a `flatten_configure_section` method has been added to the
file loader. It will convert the question tree for a particular
section into a single level hash based on the identifiers. This
way Answers do not need to change structure.

NOTE: The identifier is the name of the question node

Also removed the `data` method from `Validation::Configure` as it
is now obsolete.