Update Configurator spec to use the new Validation
The validation of `configure.yaml` was being skipped by directly
mocking the hash that is returned from the loader. However as the
configure file is now being stored as a tree, the spec needs to be
updated.

The major change is updating the question hashes to the new format
and ensuring the hash is used.