Update Configurator to use the tree of questions
The `questions_in_section` method now returns the node in the
question tree the corresponds to the section (domain, group, etc).

This allows the tree traversal method, each to run on every node of
the tree. The only other major change is the data is now stored as
the content of the node and as a OpenStruct.

NOTE: The section node is not a question, that is why each loop
has a next to skip it.