Simplify the code for asking questions
Previously the iterators that are used to ask questions where created
in an adhoc fashion. This commit simplifies the code so the questions
are created first followed by asking them and recording answers.

The Tree::Node 'node_q' still needs to be passed into ask_questions
through the enumerator. This is because question asking is going to
be dependent on the answer of the parent question.