Improve the mocking of the Node's
Previously mocking a new node would cause the nodes list to be
completely re-initialized from scratch (based on the genders file).

This meant that any nodes objects that where previously created and
possible mocked themselves where being deleted. They where then
replaced with a new (non-mocked) version of themselves.

Instead the nodes object is now maintained and the additional node
is manually added to it. The MetalArray needs to be recreated as it
is a frozen object (so you can't use pop and push).