Prevent mocking nodes that already exist
If a node already exists and is mocked again, the behaviour of
Metalware becomes undefined as two nodes with the same name exists.

It is entirely possible (depending on order defined) that:
`alces.nodes.send(name) != alces.nodes.find_by_name(name)`

Instead a check occurs that the node does not exist and errors if it
does.