Render the results from the DynamicNamespace
A tests has been added to detect the rendering error as described
in #264. Initially it the bug was believed to be caused by the
dynamic namespace not being rendered correctly. The rendering block
must be passed in manually to all MetalRecrusiveOpenStructs
(aka HashMergerDataStructure).

This commit makes the change as a rendering block should be provided.
However this does not fix the bug.