Preserve falsy-ness in the wrapper
The wrapper is always true, thus it can not wrap falsy value. The
code has been adjusted slightly so that it handles returning both
nil and false through the same condition.

This means the nil warning needs to be issued separately. This way
the shared behaviour between nil and false can be handled together.