Remove race condition when nodes are finishing
Metalware triggers the re-render of the pxelinux file by calling the
complete_hook on the pxelinux BuildMethod. However this is done in
a separate thread.

This creates a race condition between rendering of the file and
Metalware exiting when a interrupt has been received. The fix is
not to return control back to the teardown command until all the
threads have finished.

A second interrupt would trigger this the sleep loop to break and
will force exit Metalware.