Preserve the dynamic-stack when multi-threaded
Previously their was only one stack used to store the dynamic part of
the namespace. This works fine when there is only a single thread of
execution, however it breaks when it is multi-threaded.

This is because all the threads are trying to push their individual
scopes onto the same stack. As there can only be one scope the last
thread wins.

Alternatively if there are more threads than the recursion limit,
metalware will interpret it as a infinite recursion and error.

The fix is to store the stacks against individual threads. This will
prevent threads from affecting each other.

Fixes #278