Have Loader handle loading of configure.yaml data
Rather than doing this within the `Validation::Configure` class. This
both seems more logical and avoids the need to have logic in
`Validation::Configure` to sometimes load the file and sometimes accept
the configure data as input for testing purposes, even though the former
is what we always want to do in practice.

This will also make modifying the configure data before validating and
parsing it more straightforward, which is what I'm about to do.