Extract/use shared example for configure loading
Extract this and use it both when a plugin is enabled and when it is
not. Currently there is no reason why the former case should fail when
the latter won't, as an identical code path handles both cases, but this
is about to change with the addition/committing of code to handle
loading of plugin configure questions.