Handle loading of plugin configure questions
This commit adds loading of plugin configure questions and merging of
these into the main configure questions tree, prior to validation and
the Metalware configuration process starts. The code that both handles
and tests this is slightly rough at the moment, and will be refactored
in later commits.

In brief, the loading of plugin configure questions works as follows:

- All configure questions are loaded and then validated via the
`Validation::Loader` class.

- Within this class, for every enabled plugin:

   - The configure questions for that plugin are loaded.

   - For each possible configure section (`domain`, `group` etc.), a new
   question is generated internally to Metalware for whether the plugin
   should be enabled at this level; any and all plugin questions for
   this section are then set as `dependent` on this generated question.

- The resulting configure questions for the repo and every plugin,
including the generated top-level question for each section, are then
merged to form a final configure questions hash with the standard
top-level sections.

- This final merged hash is then passed to the `Validation::Configure`
class, as before, where it is validated ready to be used by the
`configure` commands.