Split up long test for plugin configure loading
This test was long and hard-to-follow, and has now been split up to
better reflect the different aspects we want to test. In the process
some no longer necessary assertions could also be removed, since there
is no way these could fail which would not be caught by the more
specific assertions.