Invert Loader tests to iterate sections at top
Rather than requiring that each test case iterates through each of the
possible configure sections to check that it works in that case, this
commit inverts this so the sections are iterated at the top level and
used to generate the tests for that section.

This makes the individual test cases simpler and should make it harder
to forget to make test cases general across the different possible
sections in future.