Handle no configure.yaml being present for plugin
If the configure.yaml file does not exist for a plugin (or is missing
certain section keys), then the plugin enabled question should still be
generated for that plugin, and this should just have no dependents for
each section (or at least for each section which is not present). This
commit implements this; previously an error would occur in this
situation.