Add minimal access to enabled plugin info for node
Within each node namespace, can now access `plugins`, which gives an
array of 'plugin namespaces' containing info related to each enabled
plugin for the node. A plugin is enabled for a node when it has both
been enabled globally for the Metalware installation, and the node has
also been configured to use it via the standard Metalware configuration
process. Plugin namespaces currently just contain the plugin's name,
with additional properties still to be added.