Handle un-configured plugin in node namespace test
It is possible that a plugin has been globally enabled for a particular
Metalware installation, but has not been configured at any level which
would apply to a particular node.

In this situation we should handle the plugin in the same was as if it
was explicitly disabled for a given node. This already occurred, due to
the saved answer for the 'plugin enabled' question being `nil`, but
adding this case to the test ensures it will continue to do so.