Add test that plugin namespace used in node namespace
This will ensure that the plugin namespace is used when accessing
`plugins` within the node namespace. It therefore means we won't need to
test the plugin namespace through the node namespace but can test it
directly, making our tests clearer and simpler, while still retaining
full coverage of this feature.