Merge and make available plugin config in namespace
This commit makes the merged plugin config available within the Plugin
namespace. This config is loaded in the same order as for the containing
node for the Plugin namespace, but using the config files nested within
the plugin's own config directory as the input.

The standard HashMergerNamespace class (as a parent for the Plugin
namespace), and a new PluginConfig hash merger class, are used to
perform this config merging.