Rename plugin enable/disable to activate/deactivate
As requested in meeting. This removes the previously overloaded meaning
of enable/disable for plugins to mean both enabling a plugin domain-wide
(which just means its questions will be included when configuring), and
enabling a plugin for a specific node (which means a Plugin namespace
for this plugin will be available within this node's namespace). The
former meaning has now been renamed to activate/deactivate.