Provide access to enabled plugins for node by name
Within a particular repo it will be a best practice to only iterate
through the enabled plugins, rather than couple the repo to any
particular plugins being available or enabled.

However, within other parts of a plugin's own config and files we want
to be able to refer to things like
`node.plugins.example.config.some_parameter` to access the plugin's own
config, therefore we wrap the enabled plugins made available in the Node
namespace in a MetalArray to support this use case.