Make build file retrieval more generic
Rather than referring to retrieving files for nodes everywhere, make the
internals of this class (in particular RetrievalProcess) just refer to a
`namespace` (a hash merger namespace with a config which specifies some
build files) and a `files_dir` (the directory within
`/var/lib/metalware/rendered/` that files will be rendered to). The
external interface of this class is now all that refers to a node, with
the new `retrieve_for_node` method.

This separation should make it more straightforward to add support for
retrieving files for other namespaces.