DRY up determining build file locations
Previously this was duplicated, as we independently determined the
location we should render the build files to and the URL they should be
retrieved from, although the latter will depend on the former. These
both happened to give consistent results at the moment, however that
might not always be the case. This commit changes the latter to
determine the URL using the former.