Refactor RetrievalProcess to not refer to repo
Rather than having RetrievalProcess be hard-coded to look for all
relative file paths within the `files` directory, have it use whichever
template directory is now passed in when it is created. This makes it
more flexible and will allow us to more easily retrieve files with
different template directories.

This change also means it makes sense to appropriately rename the
`files_dir` parameter, and the `metal_config` parameter can be removed.