Tweak Staging.template to allow returning Templater
Previously `Staging.template` required that it be called with a block,
which it would then yield the created template to. This commit retains
this behaviour while also just returning the Templater when no block is
given. This behaviour allows more flexible usage, such as in tests where
we may want to create a Templater in a `let` block for future use.