Correct Plugin files to needed format
Plugin files need to be wrapped in the same data structure as those for
the containing Node namespace to allow access in the same way when
templating; this commit uses the new `finalize_build_files` method in
the Node namespace to do this.

With this change the test for `Plugin#files` also needed to be
corrected, and while doing this it has also been made slightly more
thorough.