Support recursive hashing of arrays in WhiteListHasher
Sometimes (such as the `plugins` value in the Node namespace) a
templating namespace might contain an array value that we want to show
with `metal view`, and this value in turn contains other values which we
want to recursively hash and have included in the `metal view` output.

This commit adds support for doing this by overriding a
`recursive_array_white_list_for_hasher` within the class including the
WhiteListHasher mixin, in a similar manner to the existing methods in
this mixin intended to be overrided.