Rename nodes_in_group method to nodes_in_gender
There is a distinction between groups and genders. Genders are defined
purely by the genders file and wraps the nodeattr command. Groups on the
other hand are configured by metalware.

As the `nodes_in_group` method works for both genders and groups, it is
better named `nodes_in_gender`.

NOTE: All groups are genders but not all genders are groups