Allow cli commands to work with secondary groups
By using the Group namespace with the NodeIdentifier helper, it limited
the cli tools to work with the primary groups only.

To allow secondary groups as well, the genders file needs to be queried
directly. The result is then converted into a namespace object using the
results from `alces.nodes.find_by_name`