Replace Tree::RubyTree with a QuestionTree class
There are a few classes that depend on the behaviour of the Tree. This
makes it difficult to use if each class needs to implement there own
tree traversal depending on the purpose.

Instead this can be abstracted away in a new QuestionTree class.

NOTE: Tree::RubyTree has a few different tree traversals inbuilt, but it
isn't super obvious which one to use.