Add identifier spec for QuestionTree
This duplicates the Validation spec a bit, however I am not sure of the
future of the validation. However this spec needs to maintain a list of
identifiers that are used by the questions.

These identifiers should be returned when <question_tree>.identifiers
are called.

In the future, the build process for QuestionTree may be consolidated
from the Validation, however this can occur at a future date.