Add method to get QuestionTree identifiers
This method is fairly straight forward (with one catch). In essence, the
command does a depth_first traveral and collects all the question
identifiers.

The catch is that not all the nodes in the QuestionTree are actually
questions. The root node is special as it contains 4 child nodes that
represent each question section (:domain, :group, :node, and :local).
This means the top 5 nodes in the QuestionTree are not actually
questions.

Thus the `each` method has been modified to skip the root level and the
first level of the tree. The catch is, this only works for depth_first
traversals. Breadth_first may not use `each` (I believe anyway). This
means the `breadth_first` and `preordered_each` traversals may still
include these nodes.

As this commit does not use these other traversals, they have not been
updated. As more functionality is added to QuestionTree, these can be
updated in the future.