Add method that determines if a node is a question
The root and first level nodes are not questions. They are required to
give the tree structure into different sections.

To make it easier to determine if a node is a question, a `question?`
method has been added. It polls the nodes content to determine if the
`identifier` has been set.