Add sleep option to power (fixes #304)
This ability used to exist for the previous `power` bash script, but was
lost when we replaced this with this new version of the command written
in Ruby. This commit adds this functionality back in, via a `--sleep`
option; this option also has a default value the same as the old
default.