Add ipmi smoke tests; fix ipmi not running command
Add couple of tests for `ipmi` command, so we at least have somewhere in
the test suite which executes this command. Currently these are very
similar to and share duplicated code with the tests for `power`.

Also fix an issue flagged up when adding these, that `ipmi` previously
had multiple ways to pass it an argument: the `--command` option, used
only when running the command on a bare metal node, as well as a command
passed as the optional second argument to this command, used only when
running a command on a VM. These have been unified so only use this
argument is used as the IPMI command to run, as this seems the more
logical way to pass this.