Do not override the each method
There are certain applications in which the original each method is
required (e.g. to display the correct error structure during
validation).

This means the `each` method can not be overridden. Instead a
`filtered_each` method will be defined to take its place. This commit
removes the overridden method and defines the spec for the new filtered
method.

This filtering method will be used for all the traversals that should
only run on questions. This makes it easier to use as the user will
no longer need to remember to skip certain non-question tree nodes.