Remove the size test as the enum can not lazy calculate
Now that the filtered traversals are returning an enum, the size method
is returning nil. This indicates that the enumerator can not lazy
calculate the length and is giving up with nil.

This does not mean it is behaving incorrectly as the enumerator objects
are QuestionTrees. Instead the block given test checks if the correct
number of questions as been included.

For this reason the failing test has been removed.