Use the `identifiers` method in Repo
Repo still assumes that it is receiving a Hash of the questions when it
is actually getting a QuestionTree.

The QuestionTree object has a handy identifiers method in built. It does
not sort nor reject duplicates as it is a direct representation of the
internal tree structure's order.

The sort and reject is thus applied externally.