Rename configure_data method to question_tree
The old name configure_data is a throw back to the original
implementation where a hash of configuration data was returned.

However now a dedicated QuestionTree object is returned instead. For
this reason, :question_tree is a better name for the method.

As :configure_data is extensively used throughout metalware, its has
not been removed completely. This can be done at a future date.