Remove the Repo class
I can't quite tell what the original purpose of Repo was, however it is
now only used by the AnswerTableCreator.

The functionality it provided can be split between QuestionTree and the
AnswerTableCreator.

For this reason Repo has been removed

The sorting and uniqueness of identifiers was previously done by the Repo
class. Now that it has been removed, it needs to be done inside
AnswerTableCreator