Remove NodeIdentifier group method and references
This might be a slightly dodgy commit as it changes however metalware
works at a fundamental level. The tests are passing but there maybe some
edge cases that now fail.

Now that genders are supported this method no longer makes sense and has
to be removed. This is because not all genders have a group namespace
and thus the method could fail silently.

As a result all references to it must also be removed. This is why the
dependency check in `build` had to be modified. Now that the `nodes`
array may contain disparate nodes, the single dependency check on the
group answers no longer works.

Also the `ipmi` command need to be modified to use `options.genders`. It
redefines the `group` method however, so further changes to it are not
required.