Mock ipmi using AlcesUtils instead of manually
Most of the code concerning running a configure command to create a
group and nodes can be done directly using the AlcesUtils.

AlcesUtils has the advantage that it will update the groups cache and
genders file, so the NodeattrInterface methods do not need to be mocked
individually.

Firstly the group needs to be created with `mock_group`. This will
automatically set `alces.group` to equal the newly added group.

Then the nodes can be added using the `mock_node` method. Similarly this
will set `alces.node` to be the last mocked node.

Both of these commands will automatically update the Faked genders file
to include the group and nodes entries. By using `include AlcesUtils` in
the spec, the `NodeattrInterface` has been mocked so it copies the Faked
genders file to a temp file in the real FS. It then uses this temp file
to run `nodeattr`.

When mocking a node or group, it will automatically set the `config` and
`answer` to be blank. The config can then be mocked using the `config`
command.