Simplify mocking in Console tests
The tests for Ipmi and Power have now been adapted to use AlcesUtils;
since the tests for Console are similar and were based on these, it
makes sense to also simplify these to work in the same way. The Console
tests can also be a bit simpler however, as we don't need to mock
group-related things for Console since Console cannot be run for a
group.