Sort exceptions into user errors
Now that there is the ability to suppress `--trace` for errors, the
existing exceptions need to be sorted.

A new exception class has been added for user errors,
UserMetalwareError. The existing errors have been sorted into that
class.

A few of the errors have been removed as they are no longer being used.