Create overview.yaml for constructing the table
To allow the overview command to be easily configurable and
maintainable, the exact fields to be displayed will be stored in the
repo.

This is because the fields will likely be dependent on ERB tags which
makes it better suited to the repo.

This commit adds the loading of the file an ensures that the number of
headers match the number of fields.