Do not use mocked scopes in namespace specs
The AlcesUtils mock methods no longer set a scope. This is a pain for
this spec as it means sharing the mocked nodes becomes difficult. The
`alces` variable isn't available in a group context and hence can't be
used to pass the objects in directly.

Instead a string representation is passed in, which is then eval'ed.
This isn't ideal however as it is in the spec it shouldn't be a massive
problem.