Add method to get hostlist for a group
A required feature of the new overview command is displaying the
hostlist for a group. Previously this condensed version of the genders
file has not been required as typically a comma separated string is more
useful.

This commit adds a new NodeattrInterface method to get the hostlist
format which is then exposed in the group namespace.