Prevent Dependency from failing for the orphan group
The orphan group typically isn't configured and thus does not have an
answer file. This causes the `view-answer` command to fail as it expects
it to exist.

The fix is to make the orphan group's answer file aways optional. If the
file does exist it should go through the regular validation.

Fixes #312