Switch the data structure used by overview
Instead of having two independently managed arrays for headers and
fields, there is now a single array. Contained within this array are
objects which in turn specify the headers and values.

At this point in time, nil's are allowed for headers and fields.
Currently there is no validation that the file is correct as it is a bit
premature to implement this.

For the time being, nils are acceptable and it is up to the user to
create the file correctly.