Allow the overview.yaml file to be missing
For the time being, not all branches of the repo will contain the
overview.yaml file. To prevent them from erroring, the raw file load is
being merged into a basic hash.

This means the group will not return a nil. This can be replaced when
proper file validation is added.